pyproject.toml
[project]
name = "gemini-mcp-bridge"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "mcp<2.0",
    "httpx>=0.28.0",
]
